update to 107
Also have another go at a systmed patch. This will fall back to ConsoleKit if the system
was not booted with systemd. Thi should be similar to heftig's first patch, but I forgot
all about that and wrote one from scratch. Maybe this magically works better, who knows.

Feel free to revert all of this, I'm just commiting it as we don't have a better way of
sharing this stuff.